Create NumberSignChecker
It's a small C++ project for university. It includes functions for calculating the signs of numbers provided by the user. The program contains a loop to check sets of five numbers with an array providing information about their signs.

Initially, it welcomes you and asks for a set of five numbers. Then, the program displays the signs of the given numbers and prompts you if you would like to repeat the process. If you decline, the program informs you that it has finished and stops."